code refactorign and tests for data actor
