+++ config/config.py
@@ -5,22 +5,6 @@ import yaml
 # Enum for the dataset format
 
 
-class DatasetFormat:
-    # Both the CPP code and the data of the functions are loaded from PICKLE files
-    PICKLE = "PICKLE"
-    # We look for informations in the pickle files, if something is missing we get it from cpp files in a dynamic way
-    HYBRID = "HYBRID"
-
-    @staticmethod
-    def from_string(s):
-        if s == "PICKLE":
-            return DatasetFormat.PICKLE
-        elif s == "HYBRID":
-            return DatasetFormat.HYBRID
-        else:
-            raise ValueError("Unknown dataset format")
-
-
 @dataclass
 class TiramisuConfig:
     workspace: str = "./experiment_dir/workspace/"
@@ -29,7 +13,6 @@ class TiramisuConfig:
 
 @dataclass
 class DatasetConfig:
-    dataset_format: DatasetFormat = DatasetFormat.PICKLE
     cpps_path: str = ""
     dataset_path: str = ""
     save_path: str = ""
@@ -39,9 +22,6 @@ class DatasetConfig:
     tags: list[str] = field(default_factory=list)
 
     def __init__(self, dataset_config_dict: Dict):
-        self.dataset_format = DatasetFormat.from_string(
-            dataset_config_dict["dataset_format"]
-        )
         self.cpps_path = dataset_config_dict["cpps_path"]
         self.dataset_path = dataset_config_dict["dataset_path"]
         self.save_path = dataset_config_dict["save_path"]
@@ -80,18 +60,6 @@ class AutoSchedulerConfig:
     machine: str = "jubail"
     tiralib_config_path: str = "./config/tiralib_config.yaml"
     clean_files: bool = True
-
-    def __post_init__(self):
-        if isinstance(self.tiramisu, dict):
-            self.tiramisu = TiramisuConfig(**self.tiramisu)
-        if isinstance(self.dataset, dict):
-            self.dataset = DatasetConfig(self.dataset)
-        if isinstance(self.experiment, dict):
-            self.experiment = Experiment(**self.experiment)
-        if isinstance(self.code_deps, dict):
-            self.code_deps = CodeDeps(**self.code_deps)
-        if isinstance(self.test, dict):
-            self.test = Test(**self.test)
 
 
 def read_yaml_file(path):
