--- utils/dataset_actor/services/pickle_data_service.py
@@ -1,82 +1,0 @@
-import pickle
-import random
-
-import numpy as np
-from utils.dataset_actor.services.base_data_service import (
-    BaseDataService,
-    TiramisuProgramCache,
-)
-
-
-class PickleDataService(BaseDataService):
-    def __init__(
-        self,
-        dataset_path: str,
-        cpps_path: str,
-        path_to_save_dataset: str,
-        shuffle: bool = False,
-        seed: int = None,
-        saving_frequency: int = 10000,
-        tags: list[str] = [],
-    ):
-        super().__init__(
-            dataset_path=dataset_path,
-            path_to_save_dataset=path_to_save_dataset,
-            shuffle=shuffle,
-            seed=seed,
-            saving_frequency=saving_frequency,
-        )
-        self.cpps_path = cpps_path
-        self.cpps = {}
-        self.tags = tags
-
-        print(
-            f"reading dataset in full pkl format: dataset pkl from {self.dataset_path} and cpps pkl from {self.cpps_path}"
-        )
-
-        with open(self.dataset_path, "rb") as f:
-            self.dataset: dict[str, dict] = pickle.load(f)
-            self.function_names: list[str] = list(self.dataset.keys())
-            if self.tags:
-                print(f"Filtering dataset by tags: {self.tags}")
-                filtered_function_names: list[str] = []
-                for program in self.function_names:
-                    if any(tag in self.dataset[program]["tags"] for tag in self.tags):
-                        filtered_function_names.append(program)
-                self.function_names = filtered_function_names
-
-        with open(self.cpps_path, "rb") as f:
-            self.cpps: dict[str, str] = pickle.load(f)
-
-        # Shuffle the dataset (can be used with random sampling turned off to get a random order)
-        if self.shuffle:
-            # Set the seed if specified (for reproducibility)
-            if self.seed is not None:
-                random.seed(self.seed)
-            random.shuffle(self.function_names)
-
-        self.dataset_size = len(self.function_names)
-
-    # Returns next function name, function data, and function cpps
-    def get_next_function(self, random=False):
-        if random:
-            function_name: str = np.random.choice(self.function_names)
-        # Choose the next function sequentially
-        else:
-            function_name = self.function_names[
-                self.current_function_index % self.dataset_size
-            ]
-            self.current_function_index += 1
-
-        return (
-            function_name,
-            TiramisuProgramCache(**self.dataset[function_name]),
-            self.cpps[function_name],
-        )
-
-    def get_function_by_name(self, function_name: str):
-        return (
-            function_name,
-            TiramisuProgramCache(**self.dataset[function_name]),
-            self.cpps[function_name],
-        )
